Add Img component
Add an <Img /> component which only renders an <img> tag with the specified src and alt attributes.
The only point of this component are the propTypes, which enforce the use of an alt tag via .isRequired.

It's also a good showcase of what a component vs. a container might look like.